Admin can be added + jsonviews updated

+++ src/main/java/nl/hsleiden/service/UserService.java
@@ -1,5 +1,6 @@
 package nl.hsleiden.service;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -11,42 +12,43 @@ import nl.hsleiden.persistence.UserDAO;
  * @author Peter van Vliet
  */
 @Singleton
-public class UserService extends BaseService<User> {
+public class UserService {
 
     private final UserDAO dao;
     
     @Inject
-    public UserService(UserDAO dao)
-    {
+    public UserService(UserDAO dao) {
         this.dao = dao;
     }
     
-    public Collection<User> getAll()
-    {
+    public ArrayList<User> getAll() {
         return dao.getAll();
     }
     
     public User getUser(String email)
     {
-        return requireResult(dao.getUser(email));
+        return dao.getUser(email);
     }
     
     public void add(User user) {
         user.setRole("GUEST");
         dao.add(user);
     }
-    
+
+    public void addAdmin(User admin) {
+        admin.setRole("ADMIN");
+        dao.add(admin);
+    }
+
     public void update(String email, User newUser) {
-        // Controleren of deze gebruiker wel bestaat
         if(getUser(email) != null){
             dao.update(email, newUser);
         }
     }
     
     public void delete(String email) {
-        // Controleren of deze gebruiker wel bestaat
-        User user = getUser(email);
-        
-        dao.delete(email);
+        if(getUser(email) != null){
+            dao.delete(email);
+        }
     }
 }
